==> notebooks/04_style.ipynb <==
import os
from pathlib import Path 
import pandas as pd
import json 
from tqdm import tqdm
from langchain_groq import ChatGroq
from langchain.schema import HumanMessage, SystemMessage 
from dotenv import load_dotenv 

load_dotenv() 

CLEAN_RAW_DIR = Path('../data/processed/cleaned_raw.json')
CLEAN_SECTIONS_DIR = Path('../data/processed/cleaned_sections.json') 
STYLED_DIR = Path('../data/styled')
STYLED_DIR.mkdir(parents=True, exist_ok=True)

# load files 
with open(CLEAN_RAW_DIR, 'r', encoding='utf-8') as f: 
    cleaned_raw = json.load(f) 
    
with open(CLEAN_SECTIONS_DIR, 'r', encoding='utf-8') as f: 
    cleaned_sections = json.load(f) 

# intialize the langchain groq client 
llm = ChatGroq(
    model='qwen/qwen3-32b',
    temperature=0.7
)

def style_text(
    text, 
    style = 'Write this text in my own grammar and tone, keep meaning same'
):
    if not text.strip(): 
        return ''
    try: 
        response = llm.invoke([
            SystemMessage(content=style),
            HumanMessage(content=text) 
        ])
        return response.content.strip() 
    except Exception as e: 
        print(f'Error styling: {e}')
        return text

# style the raw texts
styled_raw_dic = {}

for fname, text in tqdm(cleaned_raw.items(), desc='Styling raw docs'): 
    styled_raw_dic[fname] = style_text(text)

styled_raw_dic 

# style sectioned texts 
styled_section_dict = {}

for fname, sections in tqdm(cleaned_sections.items(), desc='Styling sectioned docs'): 
    styled_section_dict[fname] = {}
    for section, text in sections.items(): 
        styled_section_dict[fname][section] = style_text(text) 

# save jsons 
with open(os.path.join(STYLED_DIR, 'styled_raw.json'), 'w', encoding='utf-8') as f: 
    json.dump(styled_raw_dic, f, ensure_ascii=False, indent=2)
    
with open(os.path.join(STYLED_DIR, 'styled_sections.json'), 'w', encoding='utf-8') as f: 
    json.dump(styled_section_dict, f, ensure_ascii=False, indent=2)
