==> notebooks/05_rag.ipynb <==
import os
import chromadb
from langchain_chroma import Chroma
from langchain_core.vectorstores import VectorStoreRetriever
from langchain_community.embeddings import SentenceTransformerEmbeddings
from langchain_groq import ChatGroq 
from langchain.prompts import ChatPromptTemplate 
from langchain.chains import RetrievalQA 
from dotenv import load_dotenv

load_dotenv() 

client = chromadb.PersistentClient(path='../vectorstore')
collection = client.get_collection('academic_papers') 

# initialize embeddings function for chromadb
embedding_fn = SentenceTransformerEmbeddings(
    model_name='all-MiniLM-L6-v2',
)

# load chromadb client and collection
vectorstore = Chroma(
    client=client,
    collection_name='academic_papers',
    embedding_function=embedding_fn
)

retriever = vectorstore.as_retriever(search_kwargs={'k': 5}) 

# llm
llm = ChatGroq(
    model='qwen/qwen3-32b',
    temperature=0.2
)

# prompt template
prompt_template = ChatPromptTemplate.from_messages([
    ('system', 'You are an academic assistant. Always answer in the user\'s writing style, tone, and structure.'),
    ("human", "Question: {question}\n\nContext: {context}\n\nAnswer in the user's academic style:")
])



# rag chain setup 
rqa = RetrievalQA.from_chain_type(
    llm=llm,
    retriever=retriever,
    chain_type='stuff',
    chain_type_kwargs={'prompt': prompt_template}
)

query = "Explain the role of entropy in ecosystems."
result = rqa.run(query)


print("=== Academic Answer (User Style) ===")
print(result)
